Some timer based unit test controls are added
