+++ requirements/tests.txt
@@ -1,3 +1,3 @@
 requests==2.28.0
 pytest==7.1.2
-speed_calculator==0.2.0
+speed_calculator==0.4.1
